RavenDB-21381 - Move RevisionsStorage TestingStuff to a separate file
